requirements.txt
httpx>=0.27.0
py-clob-client>=0.16.0
python-dotenv>=1.0.0
loguru>=0.7.0
scipy>=1.12.0
numpy>=1.26.0
python-dateutil>=2.9.0
schedule>=1.2.0
